Shooting menu implemented

+++ src/utils/Util.hpp
@@ -2,8 +2,8 @@
 
 namespace db
 {
-	constexpr unsigned int str2int(const char* str, int h = 0)
+	constexpr unsigned long int str2int(const char* str, int h = 0)
 	{
-		return !str[h] ? 5381 : (str2int(str, h + 1) * 33) ^ str[h];
+		return !str[h] ? 5381 : (str2int(str, h + 1) * 33ull) ^ str[h];
 	}
 }
